fix(hxy): clean up clippy lints across the workspace
Mostly mechanical: `identity_op` (1 * 1024 * 1024), `unnecessary_cast`,
`unnecessary_lazy_evaluations`, `clone_on_copy`, `useless_format`,
`manual_unwrap_or_default`, `manual_clamp`, `manual_range_contains`,
`derivable_impls`, `collapsible_if`, `cloned_ref_to_slice_refs`,
`duplicated_attributes`, plus an unused-param + dead constant cleanup.

Suppressions where the lint conflicts with the design:
- `large_enum_variant` on hxy-imhex-lang's TopItem and Stmt -- boxing
  AST variants is a deliberate non-goal.
- `too_many_arguments` on hxy-010-lang's read_array.
- `single_range_in_vec_init` on watch::vfs_sample_ranges, which
  legitimately returns a Vec<Range<u64>> (clippy's both `vec![0..len]`
  and `[0..len]` rewrites get the type wrong).
- dead-code allow on hxy-imhex-lang's NameInterner::get to keep the
  intern/lookup/get triplet symmetric.

+++ crates/hxy-core/src/cache.rs
@@ -486,7 +486,7 @@ mod tests {
         let cache = cache_with_limit_mib(20);
         let id = cache.alloc_source_id();
         let cached = CachedSource::new(cache.clone(), id, Attribution::HexView(HexViewKey(1)), inner);
-        let want = (CHUNK_SIZE_BYTES - 100) as u64..(CHUNK_SIZE_BYTES * 2 + 50) as u64;
+        let want = (CHUNK_SIZE_BYTES - 100)..(CHUNK_SIZE_BYTES * 2 + 50);
         let got = cached.read(range(want.start, want.end)).expect("read");
         assert_eq!(got, bytes[want.start as usize..want.end as usize]);
     }
